fio: 64-bit loop counters for popen_argv argv handling
popen_argv / popen_argv_pipe build argv from args_arr.size (uint64) with uint32_t
loop counters: the POSIX execvp argv-copy loops and the Windows winBuildCommandLine
(uint32_t argc + loop) truncated past 4G args. Widen to uint64_t. No behavior change
for normal argv; matches the 64-bit container sweep.

Followup to #3175 (64-bit container API sweep).

+++ src/builtin/module_builtin_fio.cpp
@@ -827,9 +827,9 @@ namespace das {
         return out;
     }
 
-    static string winBuildCommandLine ( char ** argv, uint32_t argc ) {
+    static string winBuildCommandLine ( char ** argv, uint64_t argc ) {
         string s;
-        for ( uint32_t i = 0; i < argc; ++i ) {
+        for ( uint64_t i = 0; i < argc; ++i ) {
             if ( i ) s += ' ';
             s += winArgvEscape(argv[i]);
         }
@@ -953,7 +953,7 @@ namespace das {
         // trailing NULL element.
         vector<char *> cargv;
         cargv.reserve(args_arr.size + 1);
-        for ( uint32_t i = 0; i < args_arr.size; ++i ) {
+        for ( uint64_t i = 0; i < args_arr.size; ++i ) {
             cargv.push_back(argv[i] ? argv[i] : (char *)"");
         }
         cargv.push_back(nullptr);
@@ -1137,7 +1137,7 @@ namespace das {
         }
         vector<char *> cargv;
         cargv.reserve(args_arr.size + 1);
-        for ( uint32_t i = 0; i < args_arr.size; ++i ) {
+        for ( uint64_t i = 0; i < args_arr.size; ++i ) {
             cargv.push_back(argv[i] ? argv[i] : (char *)"");
         }
         cargv.push_back(nullptr);
